requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
-e git+https://github.com/uktrade/django-staff-sso-client/@add_email_id_support#egg=authbroker_client  # via -r requirements.in
asgiref==3.2.10           # via django
certifi==2020.6.20        # via requests, sentry-sdk
chardet==3.0.4            # via requests
dj-database-url==0.5.0    # via -r requirements.in
django-crispy-forms==1.9.2  # via -r requirements.in
django-environ==0.4.5     # via -r requirements.in
django-filter==2.3.0      # via -r requirements.in
django==3.0.8             # via -r requirements.in, django-filter, django-staff-sso-client
idna==2.10                # via requests
oauthlib==3.1.0           # via requests-oauthlib
psycopg2==2.8.5           # via -r requirements.in
pytz==2020.1              # via django
requests-oauthlib==1.3.0  # via django-staff-sso-client
requests==2.24.0          # via requests-oauthlib
sentry-sdk==0.16.1        # via -r requirements.in
sqlparse==0.3.1           # via django
urllib3==1.25.9           # via requests, sentry-sdk
